refactoring : move classes related to ActivityLogger to 'org.littleshoot.proxy.extras.logging' package.
