+++ src/main/java/org/littleshoot/proxy/Launcher.java
@@ -10,8 +10,8 @@ import org.apache.commons.cli.*;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.core.config.Configurator;
 import org.jspecify.annotations.NonNull;
-import org.littleshoot.proxy.extras.ActivityLogger;
-import org.littleshoot.proxy.extras.LogFormat;
+import org.littleshoot.proxy.extras.logging.ActivityLogger;
+import org.littleshoot.proxy.extras.logging.LogFormat;
 import org.littleshoot.proxy.extras.SelfSignedMitmManager;
 import org.littleshoot.proxy.extras.SelfSignedSslEngineSource;
 import org.littleshoot.proxy.impl.ProxyUtils;
