+++ src/main/java/org/littleshoot/proxy/extras/logging/RequestHeaderField.java
@@ -1,38 +1,38 @@
-package org.littleshoot.proxy.extras;
+package org.littleshoot.proxy.extras.logging;
 
 import io.netty.handler.codec.http.HttpRequest;
 import io.netty.handler.codec.http.HttpResponse;
 import org.littleshoot.proxy.FlowContext;
 
 /**
- * Represents a response header field that should be logged.
- * This class allows logging any custom header from HTTP responses.
+ * Represents a request header field that should be logged.
+ * This class allows logging any custom header from HTTP requests.
  */
-public class ResponseHeaderField implements LogField {
+public class RequestHeaderField implements LogField {
     
     private final String headerName;
     private final String fieldName;
     private final String description;
     
     /**
-     * Creates a new response header field.
+     * Creates a new request header field.
      * @param headerName the name of the HTTP header to extract
      */
-    public ResponseHeaderField(String headerName) {
+    public RequestHeaderField(String headerName) {
         this.headerName = headerName;
-        this.fieldName = "resp_" + headerName.toLowerCase().replaceAll("[^a-z0-9]", "_");
-        this.description = "Response header: " + headerName;
+        this.fieldName = "req_" + headerName.toLowerCase().replaceAll("[^a-z0-9]", "_");
+        this.description = "Request header: " + headerName;
     }
     
     /**
-     * Creates a new response header field with custom field name.
+     * Creates a new request header field with custom field name.
      * @param headerName the name of the HTTP header to extract
      * @param fieldName the name to use for this field in logs
      */
-    public ResponseHeaderField(String headerName, String fieldName) {
+    public RequestHeaderField(String headerName, String fieldName) {
         this.headerName = headerName;
         this.fieldName = fieldName;
-        this.description = "Response header: " + headerName;
+        this.description = "Request header: " + headerName;
     }
     
     @Override
@@ -47,7 +47,7 @@ public class ResponseHeaderField implements LogField {
     
     @Override
     public String extractValue(FlowContext flowContext, HttpRequest request, HttpResponse response, long duration) {
-        String value = response.headers().get(headerName);
+        String value = request.headers().get(headerName);
         return value != null ? value : "-";
     }
     
@@ -59,7 +59,7 @@ public class ResponseHeaderField implements LogField {
     public boolean equals(Object obj) {
         if (this == obj) return true;
         if (obj == null || getClass() != obj.getClass()) return false;
-        ResponseHeaderField that = (ResponseHeaderField) obj;
+        RequestHeaderField that = (RequestHeaderField) obj;
         return headerName.equals(that.headerName);
     }
     
@@ -70,6 +70,6 @@ public class ResponseHeaderField implements LogField {
     
     @Override
     public String toString() {
-        return "ResponseHeaderField{headerName='" + headerName + "', fieldName='" + fieldName + "'}";
+        return "RequestHeaderField{headerName='" + headerName + "', fieldName='" + fieldName + "'}";
     }
 }
